backend/pyproject.toml
[tool.poetry]
name = "judge-of-character-backend"
version = "0.1.0"
description = ""
authors = ["lukewiwa <luke.wiwa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.11.1"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
social-auth-app-django = "^4.0.0"
uvicorn = "^0.11.8"
djangorestframework-camel-case = "^1.2.0"
uritemplate = "^3.0.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "0.770"
isort = "^5.5.1"
django-stubs = "^1.5.0"
djangorestframework-stubs = "^1.2.0"
flake8 = "^3.8.3"
ipdb = "^0.13.3"
pytest-django = "^3.9.0"
ipython = "^7.18.1"
django-debug-toolbar = "^2.2"

[tool.black]
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | bundle
    | migrations
    | manage
    | wsgi
    | asgi
  )/
)
'''


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
